requirements.txt
streamlit
requests
python-docx
ollama
psutil
platformdirs
langchain
sentence-transformers
faiss-cpu
docx2txt
langchain-community
